# docs/openapi.yaml
openapi: 3.0.3
...
info:
  title: Prompt Patrol API
  version: 0.5.1
  description: |
    Instructor-facing decision-support API for flagging potentially AI-generated
    short answers in software engineering courses.

    ## Authentication model

    Prompt Patrol is an **OpenID Connect Relying Party**, not an OAuth client.
    It has no third-party resource server and requests no delegated authority.
    Entra ID answers exactly one question, once per login: *who is this person?*

    Consequence: **no access token, no refresh token, and no JWT is stored
    anywhere.** The ID token is validated, its claims extracted, and the token
    discarded. See IETF `draft-ietf-oauth-browser-based-apps` §7.1
    ("Single-Domain Browser-Based Applications").

    Session state is server-side. The browser holds an opaque 256-bit random
    identifier in a cookie; that value is a database lookup key, meaningless
    outside this system.

    The authority is **tenant-scoped**
    (`login.microsoftonline.com/{tenant-id}/...`), never `/common`. With
    `/common`, any Microsoft account could complete the flow, and a personal
    account with a colliding email address could claim a provisioned slot
    before the intended user's first login. See "Provisioning" below.

    ## Authentication vs authorization — module boundary

    Separate concerns, separate modules, enforced at different layers. The
    split is by **granularity**, not preference: the dividing question is
...

...

...
    user may do.

...

...

...

...
servers:
...
paths:
...
  /api/auth/login:
    get:
      tags: [auth]
      operationId: initiateLogin
      summary: Begin the OIDC Authorization Code flow
      x-auth-enforcement: NONE # API Gateway: authorizationType = NONE
      x-browser-navigation-only: true
      description: |
        Generates `state`, `nonce`, and a PKCE `code_verifier`, persists them
        server-side keyed by a short-lived transaction id, and redirects the
        browser to Entra's `/authorize` endpoint on the **tenant-scoped**
        authority.

        **Frontend contract:** `window.location.href = '/api/auth/login'`.
        Do not call this with `fetch()` or Axios.

        Scopes requested: `openid profile email`. Deliberately **not**
        `offline_access` — no refresh token is wanted or stored.
      parameters:
        - name: returnTo
          in: query
          required: false
          description: |
            Path to return to after successful login, for deep links.

            **MUST** be validated server-side as a site-relative path: it has
            to begin with a single `/`, must not begin with `//`, and must not
            contain a scheme. Anything else is rejected with 400.

            The regex below is documentation, not enforcement. Unvalidated,
            this parameter is an open redirect on an authentication endpoint —
            a credible phishing primitive against instructors, and the most
            likely security defect in this module.
          schema:
            type: string
            pattern: '^\/(?!\/).*$'
            maxLength: 512
            example: /history?page=2
      responses:
        "302":
          description: Redirect to the Entra ID authorization endpoint.
          headers:
            Location:
              schema:
                type: string
                format: uri
                example: https://login.microsoftonline.com/{tenant-id}/oauth2/v2.0/authorize?client_id=...&response_type=code&code_challenge=...
        "400":
          $ref: "#/components/responses/BadRequest"

  # ---------------------------------------------------------------------------
  /api/auth/callback:
    get:
      tags: [auth]
      operationId: handleCallback
      summary: OIDC redirect URI — called by Entra, not by the SPA
      x-auth-enforcement: NONE
      x-called-by: Microsoft Entra ID (browser redirect)
      description: |
        Registered in the Entra app registration as a **Web** redirect URI.
        Not `SPA` — that platform type disables client secrets, and this is a
        confidential client.

        Server-side sequence:
        1. Validate `state` against the stored transaction. Mismatch → 403.
        2. Exchange `code` + `code_verifier` + client secret at Entra's
           `/token` endpoint. *(egress required)*
        3. Fetch JWKS on cache miss or `kid` rotation. *(egress required)*
        4. Validate the ID token: signature, `iss`, `aud`, `exp`, `nonce`.
           Reject if `tid` is not the configured SMU tenant.
        5. Extract `oid`, `tid`, `email`, `name`, `sid`, `login_hint`.
        6. **Delegate to the users module** —
           `users.resolve_or_reject(oid, tid, email)`. See below.
        7. Insert a fresh `sessions` row — a **new** session id on every login,
           which defeats session fixation.
        8. **Discard the ID token.** It is not persisted.

        Steps 2 and 3 are the *only* backend-to-Entra traffic in the entire
        system, and the sole justification for the NAT Gateway.

        ### Step 6 — provisioning resolution
        Auth cannot decide whether a person may use this system; that is a
        users-module policy. But auth needs a `users.id` before it can insert a
        session row. It therefore **asks**:

        ```python
        user = await users.resolve_or_reject(oid=..., tid=..., email=...)
        ```

        The users module resolves in this order:

        | Case | Match | Action |
        |------|-------|--------|
        | Returning user | `(tenant_id, external_id)`, `deleted_at IS NULL` | Return the row |
        | First login | `email` (case-insensitive) where `external_id IS NULL` and `deleted_at IS NULL` | Bind `oid`/`tid`, set `display_name`, stamp `first_login_at`, return |
        | Deactivated | match found (by either rule above) but `deleted_at IS NOT NULL` | **Reject** — `deactivated`, not `not_provisioned` |
        | Unprovisioned | no match at all | **Reject** — `not_provisioned` |

        On rejection: **no user row is created/modified and no session is
        issued.** Redirect to the SPA's `not_provisioned` or `deactivated`
        route accordingly — these are deliberately distinguishable in the
        response and in the audit log (never had access, vs. access was
        revoked), even if you choose to show the end user the same generic
        front-end copy for both. Repeated unprovisioned or deactivated login
        attempts against this endpoint are a signal worth retaining.

        Auth never names a role, never writes `users.role`, and never decides
        eligibility. It delegates and acts on the answer.
      parameters:
        - name: code
          in: query
          schema: { type: string }
          description: Authorization code. Present on success.
        - name: state
          in: query
          required: true
          schema: { type: string }
        - name: error
          in: query
          schema: { type: string }
          description: Present instead of `code` when Entra rejects the request.
          example: access_denied
        - name: error_description
          in: query
          schema: { type: string }
      responses:
        "302":
          description: |
            On success: session cookie set, redirect to the SPA (to `returnTo`
            if supplied and validated, else `/`).

            On rejected provisioning: **no cookie**, redirect to the SPA
            `not_provisioned` route.

            On an Entra-side error: redirect to the SPA error route with a
            safe, non-reflected error code. Never render `error_description`
            back to the browser unescaped.
          headers:
            Set-Cookie:
              schema:
                type: string
                example: __Host-pp_session=<opaque>; Path=/; Secure; HttpOnly; SameSite=Strict; Max-Age=14400
            Location:
              schema: { type: string, format: uri }
        "403":
          description: |
            `state` mismatch, missing, or expired, or `tid` is not the
            configured tenant. Indicates CSRF against the redirect URI, a stale
            browser tab, or a cross-tenant attempt. Logged as a security event.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "502":
          description: Entra token endpoint unreachable or returned an error.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  # ---------------------------------------------------------------------------
...
  /api/auth/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: Revoke all sessions for the current user
      security:
        - sessionCookie: []
      x-auth-enforcement: GATEWAY
      description: |
        Global logout. Revokes **every** active session for this `user_id`,
        not just the calling browser's — the chosen semantics for an academic
        integrity tool.

        Requires no role check: a user may always terminate their own sessions.

        Sessions are soft-deleted (`revoked_at = now()`), never hard-deleted,
        preserving referential integrity with audit entries.

        Idempotent: repeating it with a token whose session is already revoked
        revokes nothing further and still completes. The caller is resolved
        through a **live** session only — otherwise a copied cookie would stay
        usable indefinitely as a "sign this person out everywhere" primitive.

        ### Two-step logout — the server completes it
        Revoking the local sessions is not sufficient. The Entra SSO session
        survives, so the user can click "log in" and be silently readmitted.
        This looks like a logout bug and is not.

        **Implemented as a redirect, not a header (corrected 2026-08-17).**
        This endpoint answers `303` straight to the Entra sign-out URL, and
        `SignOutForm.tsx` submits a native form POST rather than an XHR.

        The originally specified shape — `204` plus an `X-Entra-Logout-Url`
        header for the SPA to follow in JavaScript — was **not** built, on
        purpose. The full document navigation is what discards the in-memory
        query cache, and the redirect works with no JavaScript at all. The
        implementation is the better design; this text was corrected to match
        it rather than the reverse.

        Either way the URL is built server-side, so tenant configuration never
        enters the frontend bundle and does not vary between environments.

        ### On `logout_hint`
        The URL includes `logout_hint` when available, suppressing Entra's
        account-picker prompt. Its value is the `login_hint` **claim issued by
        Entra**, captured at callback.

        It is *not* `users.id` and *not* the UPN. Both are ignored by Entra —
        `users.id` is a UUID this system mints that Entra has never seen, and
        Microsoft's documentation explicitly warns against passing UPNs. Either
        substitution silently produces the account picker.

        Requires the `login_hint` optional claim to be enabled on the ID token
        in the Entra app registration (Token configuration).

        ### Residual behaviour — document, do not fix
        On domain-joined machines holding a valid Primary Refresh Token, the
        next sign-in may complete silently even after a correct
        `end_session_endpoint` redirect. That is device SSO working as
        designed. Note it in the test plan so nobody spends a day on it.
      responses:
        "303":
          description: |
            All live sessions for this user revoked; cookie cleared; browser
            redirected onward to Entra sign-out.

            Returned even when the cookie is absent or already revoked — the
            endpoint never reports whether a session existed, and the user ends
            up signed out of Entra either way.
          headers:
            Location:
              required: true
              schema:
                type: string
                format: uri
                example: https://login.microsoftonline.com/{tenant-id}/oauth2/v2.0/logout?post_logout_redirect_uri=https%3A%2F%2Fpromptpatrol.smu.edu.sg&logout_hint=...
              description: |
                `post_logout_redirect_uri` must be registered in the Entra app
                registration, or Entra silently ignores it and strands the user
                on a generic Microsoft page. Register production and
                `http://localhost:5173` for development.

                Falls back to `{FRONTEND_URL}/login` if the discovery document
                cannot be reached — the local sessions are revoked regardless.
            Set-Cookie:
              schema:
                type: string
                example: __Host-session=; Path=/; Secure; HttpOnly; SameSite=Strict; Max-Age=0

  # ---------------------------------------------------------------------------
  # ADMIN — BLOCKED ON USERS MODULE
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-pp_session
      description: |
        Opaque 256-bit CSPRNG session identifier. Carries no claims and is not
        a token — it is a primary key into the `sessions` table and is
        meaningless outside this system.

        Attributes (per `draft-ietf-oauth-browser-based-apps` §6.1.3.2):
        `Secure` and `HttpOnly` are MUST; `SameSite=Strict`, `Path=/`, absent
        `Domain`, and the `__Host-` prefix are SHOULD and all applied.

        Establishes identity only. Authorization is evaluated separately by the
        users module.
  schemas:
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Stable machine-readable code. Safe to branch on.
          enum:
            - invalid_request
            - unauthenticated # 401 — no valid session (auth, gateway)
            - session_expired # 401 — session existed, clock fired (auth)
            - not_provisioned # authenticated at Entra, no user record (users)
            - deactivated # authenticated at Entra, user record exists
              # but soft-deleted (users)
            - forbidden # 403 — valid session, insufficient role (users)
            - not_found
            - conflict # 409 — duplicate email, or invalid state
              # transition e.g. restoring a non-deleted user (users)
            - state_mismatch
            - upstream_idp_error
            - payload_too_large # 413 — answer_text over 10,000 chars (checks)
            - question_text_required # 400 — served detector scores on
              # (question, answer) pairs and question_text
              # was omitted. See GET /api/detector. (checks)
            - detector_unavailable # 503 — detector service down (checks)
            - detector_timeout # 504 — detector exceeded 10s budget (checks)
            - internal_error
        message:
          type: string
          description: |
            Human-readable. Never interpolates unsanitised upstream input
            (notably Entra's `error_description`).
        request_id:
          type: string
          description: Correlates with CloudWatch logs and the audit trail.
  responses:
    BadRequest:
      description: Malformed or invalid request.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...